+++ fractal_tasks_core/tasks/image_based_registration_hcs_init.py
@@ -17,7 +17,7 @@ from typing import Any
 
 from pydantic import validate_call
 
-from fractal_tasks_core.tasks._registration_utils import (
+from fractal_tasks_core.utils import (
     create_well_acquisition_dict,
 )